Use block syntax for :set
If you don't use the block syntax for `:set`, the listed variable are evaluated in the current context, rather than in the context in which foreman is run - ex. this way you can have the foreman defaults set in one place but have different settings for `shared_path` based on the environment.